requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
anyio==4.3.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.8.1
asttokens==2.4.1
async-lru==2.0.4
attrs==23.2.0
Babel==2.14.0
backoff==2.2.1
beautifulsoup4==4.12.3
bleach==6.1.0
certifi==2024.8.30
cffi==1.16.0
charset-normalizer==3.3.2
colorama==0.4.6
comm==0.2.2
contourpy==1.2.1
cycler==0.12.1
debugpy==1.8.1
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.8
Django==5.0.7
django-allauth==0.63.3
django-background-tasks==1.2.8
django-environ==0.11.2
django-shopify-auth==2.0.1
djangorestframework==3.15.2
ecdsa==0.19.0
executing==2.0.1
fastjsonschema==2.19.1
filelock==3.15.4
fonttools==4.53.0
fqdn==1.5.1
frozenlist==1.4.1
geographiclib==2.0
geopy==2.4.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.6
ipykernel==6.29.4
ipython==8.23.0
ipywidgets==8.1.2
isoduration==20.11.0
jedi==0.19.1
Jinja2==3.1.3
json5==0.9.24
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.10.0
jupyter-lsp==2.2.4
jupyter_client==8.6.1
jupyter_core==5.7.2
jupyter_server==2.13.0
jupyter_server_terminals==0.5.3
jupyterlab==4.1.5
jupyterlab_pygments==0.3.0
jupyterlab_server==2.25.4
jupyterlab_widgets==3.0.10
kiwisolver==1.4.5
MarkupSafe==2.1.5
matplotlib==3.9.0
matplotlib-inline==0.1.6
mistune==3.0.2
multidict==6.0.5
mysqlclient==2.2.4
nbclient==0.10.0
nbconvert==7.16.3
nbformat==5.10.3
nest-asyncio==1.6.0
notebook==7.1.2
notebook_shim==0.2.4
numpy==2.0.0
opencage==3.0.2
overrides==7.7.0
packaging==24.0
pandocfilters==1.5.1
parso==0.8.3
phonenumbers==8.13.47
pillow==10.3.0
platformdirs==4.2.0
prometheus_client==0.20.0
prompt-toolkit==3.0.43
psutil==5.9.8
psycopg2==2.9.9
pure-eval==0.2.2
pyactiveresource==2.2.2
pyasn1==0.6.0
pycparser==2.22
Pygments==2.17.2
PyJWT==2.8.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.1
python-jose==3.3.0
python-json-logger==2.0.7
pywin32==306
pywinpty==2.0.13
PyYAML==6.0.1
pyzmq==25.1.2
qtconsole==5.5.1
QtPy==2.4.1
referencing==0.34.0
requests==2.31.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.18.0
rsa==4.9
Send2Trash==1.8.2
setuptools==70.1.0
ShopifyAPI==12.5.0
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
sqlparse==0.5.0
stack-data==0.6.3
terminado==0.18.1
tinycss2==1.2.1
tornado==6.4
tqdm==4.66.5
traitlets==5.14.2
types-python-dateutil==2.9.0.20240316
tzdata==2024.1
uri-template==1.3.0
urllib3==2.2.2
virtualenv==20.26.3
wcwidth==0.2.13
webcolors==1.13
webencodings==0.5.1
websocket-client==1.7.0
wheel==0.44.0
widgetsnbextension==4.0.10
yarl==1.11.0
